requirements.txt
asgiref==3.6.0
Brotli==1.0.9
dj-database-url==1.2.0
Django==3.2.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.4.0
